post remover only for archived

+++ src/models/post/PostRemover.php
@@ -11,7 +11,6 @@ use bizley\podium\api\models\thread\Thread;
 use bizley\podium\api\models\thread\ThreadRemover;
 use bizley\podium\api\repos\PostRepo;
 use Yii;
-use yii\db\Exception;
 
 /**
  * Class PostRemover
@@ -49,37 +48,33 @@ class PostRemover extends PostRepo implements RemovableInterface
         if (!$this->beforeRemove()) {
             return false;
         }
-        $transaction = Yii::$app->db->beginTransaction();
+        if (!$this->archived) {
+            $this->addError('archived', Yii::t('podium.error', 'post.must.be.archived'));
+            return false;
+        }
+
         try {
+            $thread = $this->getThreadModel();
+            if ($thread->posts_count === 0 && $thread->archived) {
+                if (!$thread->convert(ThreadRemover::class)->remove()) {
+                    Yii::error('Error while deleting empty archived thread', 'podium');
+                    return false;
+                }
+
+                $this->afterRemove();
+                return true;
+            }
+
             if ($this->delete() === false) {
                 Yii::error('Error while deleting post', 'podium');
-                throw new Exception('Error while deleting post!');
-            }
-
-            $thread = $this->getThreadModel();
-
-            if (!$thread->updateCounters(['posts_count' => -1])) {
-                throw new Exception('Error while updating thread counters!');
-            }
-            if (!$thread->getParent()->updateCounters(['posts_count' => -1])) {
-                throw new Exception('Error while updating forum counters!');
-            }
-            if ($thread->posts_count === 0 && !$thread->convert(ThreadRemover::class)->remove()) {
-                throw new Exception('Error while removing empty thread!');
+                return false;
             }
 
             $this->afterRemove();
-
-            $transaction->commit();
             return true;
 
         } catch (\Throwable $exc) {
             Yii::error(['Exception while removing post', $exc->getMessage(), $exc->getTraceAsString()], 'podium');
-            try {
-                $transaction->rollBack();
-            } catch (\Throwable $excTrans) {
-                Yii::error(['Exception while post removing transaction rollback', $excTrans->getMessage(), $excTrans->getTraceAsString()], 'podium');
-            }
         }
         return false;
     }
